refactor: Changes in INV4
